backend/requirements.txt
Django==3.0.6
djangorestframework==3.11.0
gunicorn==19.9.0
whitenoise==5.1.0

